Refactoring into smaller submodules
